Move EntropyRng to a new module
